backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-socketio==5.11.3
python-engineio==4.9.1
pydantic==2.7.1
